📦 Struct: Apart controls.
