+++ kitx_mobile/lib/pages/routes.dart
@@ -1,7 +1,4 @@
-﻿// ignore_for_file: non_constant_identifier_names
-
-import 'package:get/get.dart';
-
+﻿import 'package:get/get.dart';
 import 'package:kitx_mobile/pages/about_page.dart';
 import 'package:kitx_mobile/pages/account_page.dart';
 import 'package:kitx_mobile/pages/device_page.dart';
@@ -10,30 +7,11 @@ import 'package:kitx_mobile/pages/settings_page.dart';
 import 'package:kitx_mobile/pages/test_page.dart';
 
 /// Get Pages
-List<GetPage<dynamic>> GetPages() => [
-      // 定义路由
-      GetPage(
-        name: '/',
-        page: () => HomePage(),
-      ),
-      GetPage(
-        name: '/DevicePage/',
-        page: () => DevicePage(),
-      ),
-      GetPage(
-        name: '/AccountPage/',
-        page: () => AccountPage(),
-      ),
-      GetPage(
-        name: '/TestPage/',
-        page: () => TestPage(),
-      ),
-      GetPage(
-        name: '/SettingsPage/',
-        page: () => SettingsPage(),
-      ),
-      GetPage(
-        name: '/AboutPage/',
-        page: () => AboutPage(),
-      ),
+List<GetPage<dynamic>> getPages() => [
+      GetPage(name: '/', page: () => HomePage()),
+      GetPage(name: '/DevicePage/', page: () => DevicePage()),
+      GetPage(name: '/AccountPage/', page: () => AccountPage()),
+      GetPage(name: '/TestPage/', page: () => TestPage()),
+      GetPage(name: '/SettingsPage/', page: () => SettingsPage()),
+      GetPage(name: '/AboutPage/', page: () => AboutPage()),
     ];
